[DOP-15650] Rename permissions endpoint
