+++ docs/conf.py
@@ -34,7 +34,7 @@ author = "DataOps.ETL"
 # The short X.Y version.
 
 # this value is updated automatically by `poetry version ...` and poetry-bumpversion plugin
-ver = Version.parse("0.1.4")
+ver = Version.parse("0.2.0")
 version = ver.base_version
 # The full version, including alpha/beta/rc tags.
 release = ver.public
