+++ horizon/backend/api/v1/router/__init__.py
@@ -9,7 +9,6 @@ from horizon.backend.api.v1.router.namespace_history import (
     router as namespace_history_router,
 )
 from horizon.backend.api.v1.router.namespaces import router as namespaces_router
-from horizon.backend.api.v1.router.permission import router as permission_router
 from horizon.backend.api.v1.router.users import router as users_router
 
 router = APIRouter(prefix="/v1")
@@ -19,4 +18,3 @@ router.include_router(namespaces_router)
 router.include_router(hwm_router)
 router.include_router(hwm_history_router)
 router.include_router(namespace_history_router)
-router.include_router(permission_router)
